requirements.txt
Pillow==7.1.0
pytesseract==0.3.3
Wand==0.5.9
